requirements.txt
python-telegram-bot
ccxt
python-dotenv